Still refactoring code and splitting in to more files. Still need to refactor mlx_handler.c file into more functions and fix texture implementation
